+++ yggdrasil_decision_forests/learner/distributed_decision_tree/dataset_cache/dataset_cache_reader_test.cc
@@ -20,6 +20,10 @@
 #include "absl/log/check.h"
 #include "absl/log/log.h"
 #include "absl/strings/str_cat.h"
+#include "absl/strings/string_view.h"
+#include "absl/types/optional.h"
+#include "yggdrasil_decision_forests/dataset/data_spec.h"
+#include "yggdrasil_decision_forests/dataset/data_spec.pb.h"
 #include "yggdrasil_decision_forests/dataset/data_spec_inference.h"
 #include "yggdrasil_decision_forests/dataset/vertical_dataset.h"
 #include "yggdrasil_decision_forests/dataset/vertical_dataset_io.h"
@@ -41,14 +45,18 @@ using test::StatusIs;
 
 class End2End : public ::testing::Test {
  public:
-  void SetUp() override {
+  void Prepare(absl::string_view dataset_name, absl::string_view label_key,
+               absl::optional<absl::string_view> group_column) {
     // Prepare the dataspec.
     const auto dataset_path = absl::StrCat(
         "csv:",
-        file::JoinPath(test::DataRootDirectory(),
-                       "yggdrasil_decision_forests/test_data/"
-                       "dataset/adult_train.csv"));
-    dataset::CreateDataSpec(dataset_path, false, {}, &data_spec_);
+        file::JoinPath(
+            test::DataRootDirectory(),
+            absl::StrCat("yggdrasil_decision_forests/test_data/"
+                         "dataset/",
+                         dataset_name, "_train.csv")));
+    dataset::CreateDataSpec(dataset_path, false, guide_, &data_spec_);
+    LOG(INFO) << "Dataspec:" << dataset::PrintHumanReadable(data_spec_);
 
     // Shard the dataset.
     CHECK_OK(LoadVerticalDataset(dataset_path, data_spec_, &dataset_));
@@ -69,11 +77,19 @@ class End2End : public ::testing::Test {
     config.set_max_unique_values_for_discretized_numerical(32);
 
     int32_t label_column_idx;
-    CHECK_OK(dataset::GetSingleColumnIdxFromName("income", data_spec_,
+    CHECK_OK(dataset::GetSingleColumnIdxFromName(label_key, data_spec_,
                                                  &label_column_idx));
     config.set_label_column_idx(label_column_idx);
 
-    cache_path_ = file::JoinPath(test::TmpDirectory(), "cache");
+    if (group_column.has_value()) {
+      int32_t group_column_idx;
+      CHECK_OK(dataset::GetSingleColumnIdxFromName(
+          group_column.value(), data_spec_, &group_column_idx));
+      config.set_group_column_idx(group_column_idx);
+    }
+
+    cache_path_ = file::JoinPath(test::TmpDirectory(),
+                                 absl::StrCat(dataset_name, "_cache"));
 
     EXPECT_OK(CreateDatasetCacheFromShardedFiles(sharded_dataset_path,
                                                  data_spec_, {}, cache_path_,
@@ -83,9 +99,12 @@ class End2End : public ::testing::Test {
   dataset::proto::DataSpecification data_spec_;
   std::string cache_path_;
   dataset::VerticalDataset dataset_;
+  dataset::proto::DataSpecificationGuide guide_;
 };
 
-TEST_F(End2End, Base) {
+TEST_F(End2End, Adult) {
+  Prepare("adult", "income", {});
+
   proto::DatasetCacheReaderOptions options;
   auto reader = DatasetCacheReader::Create(cache_path_, options).value();
   EXPECT_EQ(reader->num_examples(), 22792);
@@ -157,6 +176,37 @@ TEST_F(End2End, Base) {
     EXPECT_EQ(count, reader->num_examples());
     EXPECT_EQ(reader->DiscretizedNumericalFeatureBoundaries(4).size(), 15);
   }
+
+  LOG(INFO) << reader->MetadataInformation();
+}
+
+TEST_F(End2End, SyntheticRanking) {
+  auto* hash_column = guide_.add_column_guides();
+  hash_column->set_column_name_pattern("^GROUP$");
+  hash_column->set_type(dataset::proto::ColumnType::HASH);
+
+  auto* cat_column = guide_.add_column_guides();
+  cat_column->set_column_name_pattern("^cat_.*");
+  cat_column->set_type(dataset::proto::ColumnType::CATEGORICAL);
+
+  Prepare("synthetic_ranking", "LABEL", "GROUP");
+
+  proto::DatasetCacheReaderOptions options;
+  options.add_features(6);
+  options.add_features(2);
+  options.set_load_all_features(false);
+  auto reader = DatasetCacheReader::Create(cache_path_, options).value();
+  EXPECT_EQ(reader->num_examples(), 3990);
+  EXPECT_EQ(reader->ranking_labels().size(), reader->num_examples());
+  EXPECT_EQ(reader->ranking_groups().size(), reader->num_examples());
+
+  // The label and hash columns are not loaded in memory.
+  EXPECT_THAT(reader->InOrderHashFeatureValueIterator(0).status(),
+              StatusIs(absl::StatusCode::kInvalidArgument,
+                       "Column 0 is not available"));
+  EXPECT_THAT(reader->InOrderNumericalFeatureValueIterator(1).status(),
+              StatusIs(absl::StatusCode::kInvalidArgument,
+                       "Column 1 is not available"));
 
   LOG(INFO) << reader->MetadataInformation();
 }