+++ yggdrasil_decision_forests/learner/distributed_gradient_boosted_trees/distributed_gradient_boosted_trees.cc
@@ -332,6 +332,9 @@ absl::Status CreateDatasetCacheFromPartialDatasetCache(
     const dataset::proto::DataSpecification& data_spec) {
   auto create_cache_config = spe_config.create_cache();
   create_cache_config.set_label_column_idx(config_link.label());
+  if (config_link.has_ranking_group() && config_link.ranking_group() >= 0) {
+    create_cache_config.set_group_column_idx(config_link.ranking_group());
+  }
   if (config_link.has_weight_definition()) {
     if (!config_link.weight_definition().has_numerical()) {
       return absl::InvalidArgumentError(
@@ -356,6 +359,9 @@ absl::Status CreateDatasetCache(
     const dataset::proto::DataSpecification& data_spec) {
   auto create_cache_config = spe_config.create_cache();
   create_cache_config.set_label_column_idx(config_link.label());
+  if (config_link.has_ranking_group() && config_link.ranking_group() >= 0) {
+    create_cache_config.set_group_column_idx(config_link.ranking_group());
+  }
   if (config_link.has_weight_definition()) {
     if (!config_link.weight_definition().has_numerical()) {
       return absl::InvalidArgumentError(