+++ yggdrasil_decision_forests/utils/own_or_borrow.h
@@ -58,6 +58,17 @@ class VectorOwnOrBorrow {
 
   size_t size() const { return values_.size(); }
 
+  // Is the data empty.
+  bool empty() const { return values_.empty(); }
+
+  // Releases the data. If the data was owned, the data is cleared (e.g.,
+  // released from memory). If the data was borrowed, the data is released.
+  void release() {
+    owned_values_.clear();
+    values_ = owned_values_;
+    owner_ = true;
+  }
+
  private:
   absl::Span<const T> values_;
   std::vector<T> owned_values_;