+++ CHANGELOG.md
@@ -8,9 +8,11 @@ Changelog under `yggdrasil_decision_forests/port/python/CHANGELOG.md`.
 ### Features
 
 -   Speed-up training of GBT models by ~10%
--   Rename LAMBDA_MART_NDCG5 to LAMBDA_MART_NDCG. The old name is deprecated
-    but can still be used.
+-   Rename LAMBDA_MART_NDCG5 to LAMBDA_MART_NDCG. The old name is deprecated but
+    can still be used.
 -   Allow configuring the truncation of NDCG losses.
+-   Add support for distributed training for ranking gradient boosted tree
+    models.
 
 ### Misc