+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -25,8 +25,8 @@
 -   Speed-up training of GBT models by ~10%
 -   Add `ydf.util.read_tf_record` and `ydf.util.write_tf_record` to facilitate
     TF Record datasets usage.
--   Rename LAMBDA_MART_NDCG5 to LAMBDA_MART_NDCG. The old name is deprecated
-    but can still be used.
+-   Rename LAMBDA_MART_NDCG5 to LAMBDA_MART_NDCG. The old name is deprecated but
+    can still be used.
 -   Allow configuring the truncation of NDCG losses.
 -   Enable multi-threading when using `model.predict` and `model.evaluate`.
 -   Default number of threads of `model.analyze` is equal to the number of
@@ -37,6 +37,8 @@
     same way as categorical values, and validate their types.
 -   Native support for PyGrain DataLoader and Dataset for all operations (e.g.,
     training, evaluation, predictions).
+-   Add support for distributed training for ranking gradient boosted tree
+    models.
 
 ### Fix